+++ deltafi-ui/src/mocks/graphql/getEgressFlowByName.js
@@ -16,64 +16,86 @@
    limitations under the License.
 */
 
+const errors = [
+  {
+    configName: "passthroughv2",
+    errorType: "UNRESOLVED_VARIABLE",
+    message: "Could not resolve placeholder 'notFound' in value \"${notFound}\"",
+  },
+  {
+    configName: "PassthroughLoadAction",
+    errorType: "INVALID_ACTION_PARAMETERS",
+    message: "$.unexpected: is not defined in the schema and the schema does not allow additional properties",
+  },
+  {
+    configName: "PassthroughTransformAction",
+    errorType: "UNREGISTERED_ACTION",
+    message: "Action: org.deltafi.passthrough.action.MissingRoteTransformAction has not been registered with the system",
+  },
+  {
+    configName: "NotReallyTranformAction",
+    errorType: "INVALID_ACTION_CONFIG",
+    message: "Action: org.deltafi.passthrough.action.RoteLoadAction is not registered as a TransformAction",
+  },
+];
+
+const flowStatus = ["STOPPED", "ERRORS"];
+const flowStatusMap = new Map([
+  ["STOPPED", { state: "STOPPED", errors: [] }],
+  ["ERRORS", { state: "STOPPED", errors: errors }],
+]);
+
 const generateData = () => {
   return {
-      "name": "passthrough",
-      "description": "Egress flow that passes data through unchanged",
-      "sourcePlugin": {
-        "groupId": "org.deltafi.passthrough",
-        "artifactId": "deltafi-passthrough",
-        "version": "0.95.4"
+    name: "passthrough",
+    description: "Egress flow that passes data through unchanged",
+    sourcePlugin: {
+      groupId: "org.deltafi.passthrough",
+      artifactId: "deltafi-passthrough",
+      version: "0.95.4",
+    },
+    flowStatus: flowStatusMap.get(flowStatus[Math.floor(Math.random() * flowStatus.length)]),
+    includeIngressFlows: ["passthrough"],
+    excludeIngressFlows: [],
+    formatAction: {
+      name: "passthrough.PassthroughFormatAction",
+      parameters: null,
+      apiVersion: null,
+      requiresDomains: ["binary"],
+      requiresEnrichment: [],
+      type: "org.deltafi.passthrough.action.RoteFormatAction",
+    },
+    validateActions: [
+      {
+        name: "passthrough.PassthroughValidateAction",
+        apiVersion: null,
+        parameters: null,
+        type: "org.deltafi.passthrough.action.RubberStampValidateAction",
       },
-      "flowStatus": {
-        "state": "RUNNING",
-        "errors": []
+    ],
+    egressAction: {
+      apiVersion: null,
+      name: "passthrough.PassthroughEgressAction",
+      parameters: {
+        egressFlow: "egressFlow",
+        metadataKey: "deltafiMetadata",
+        url: "http://deltafi-egress-sink-service",
       },
-      "includeIngressFlows": [
-        "passthrough"
-      ],
-      "excludeIngressFlows": [],
-      "formatAction": {
-        "name": "passthrough.PassthroughFormatAction",
-        "parameters": null,
-        "apiVersion": null,
-        "requiresDomains": [
-          "binary"
-        ],
-        "requiresEnrichment": [],
-        "type": "org.deltafi.passthrough.action.RoteFormatAction"
+      type: "org.deltafi.core.action.RestPostEgressAction",
+    },
+    variables: [
+      {
+        name: "egressUrl",
+        value: null,
+        dataType: "STRING",
+        description: "The URL to post the DeltaFile to",
+        defaultValue: "http://deltafi-egress-sink-service",
+        required: true,
       },
-      "validateActions": [
-        {
-          "name": "passthrough.PassthroughValidateAction",
-          "apiVersion": null,
-          "parameters": null,
-          "type": "org.deltafi.passthrough.action.RubberStampValidateAction"
-        }
-      ],
-      "egressAction": {
-        "apiVersion": null,
-        "name": "passthrough.PassthroughEgressAction",
-        "parameters": {
-          "egressFlow": "egressFlow",
-          "metadataKey": "deltafiMetadata",
-          "url": "http://deltafi-egress-sink-service"
-        },
-        "type": "org.deltafi.core.action.RestPostEgressAction"
-      },
-      "variables": [
-        {
-          "name": "egressUrl",
-          "value": null,
-          "dataType": "STRING",
-          "description": "The URL to post the DeltaFile to",
-          "defaultValue": "http://deltafi-egress-sink-service",
-          "required": true
-        }
-      ]
-    }
+    ],
+  };
 };
 
 export default {
-  "getEgressFlow": generateData()
-}
+  getEgressFlow: generateData(),
+};